Factor out min payment amount logic
